all functions moved to config.cpp
