+++ src/MF_CustomDevice/CustomDevice.h
@@ -8,5 +8,4 @@ namespace CustomDevice
     void update();
     void OnSet();
     void PowerSave(bool state);
-    void GetConfigFromFlash();
 }
